docker/streamlit/requirements.txt
streamlit
pymongo
python-dotenv
plotly